// src/shared/dataprovider/api/model/cashInOutStatusEnum.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type CashInOutStatusEnum = 1 | 2 | 3 | 4 | 5;

export const CashInOutStatusEnum = {
    NUMBER_1: 1 as CashInOutStatusEnum,
    NUMBER_2: 2 as CashInOutStatusEnum,
    NUMBER_3: 3 as CashInOutStatusEnum,
    NUMBER_4: 4 as CashInOutStatusEnum,
    NUMBER_5: 5 as CashInOutStatusEnum
};