// src/shared/dataprovider/api/model/transactionTypeEnum.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type TransactionTypeEnum = 1 | 2;

export const TransactionTypeEnum = {
    NUMBER_1: 1 as TransactionTypeEnum,
    NUMBER_2: 2 as TransactionTypeEnum
};