// src/shared/dataprovider/api/model/agentListData.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentStatus } from './agentStatus';
import { AgentTypeEnum } from './agentTypeEnum';
import { BrandDTO } from './brandDTO';
import { SubAgentDTO } from './subAgentDTO';

export interface AgentListData { 
    AgentId?: number;
    AgentName?: string;
    Username?: string;
    Email?: string;
    AgentType?: AgentTypeEnum;
    Status?: AgentStatus;
    LastLogin?: Date;
    DateCreated?: Date;
    DownLineCount?: number;
    ClientCount?: number;
    WalletBalance?: number;
    RemittanceBalance?: number;
    TotalCiCommission?: number;
    TotalCoCommission?: number;
    BankWithdrawalFee?: number;
    USDTWithdrawalFee?: number;
    CashPickUpWithdrawalFee?: number;
    Brands?: Array<BrandDTO>;
    SubAgents?: Array<SubAgentDTO>;
    UserId? : number;
}