// src/shared/dataprovider/api/model/agentReportTransactionLogList.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CashInOutStatusEnum } from './cashInOutStatusEnum';
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface AgentReportTransactionLogList { 
    Timestamp?: Date;
    TransactionNo?: string;
    ClientName?: string;
    Type?: TransactionTypeEnum;
    VendorName?: string;
    Channelname?: string;
    GrossAmount?: number;
    Commissions?: number;
    NetAmount?: number;
    Status?: CashInOutStatusEnum;
}