// src/shared/dataprovider/api/model/brandDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StatusEnum } from './statusEnum';

export interface BrandDTO { 
    ClientId?: number;
    ClientName?: string;
    DateAdded?: Date;
    WalletBalance?: number;
    CiCommission?: number;
    CiCount?: number;
    CiTotal?: number;
    CoCommission?: number;
    CoCount?: number;
    CoTotal?: number;
    TotalCommission?: number;
    Status?: StatusEnum;
    SubBrands?: Array<BrandDTO>;
    EffectivityDate? : string;
}